refactor: reorganize imports and update type definitions across multiple files
