+++ CHANGELOG.md
@@ -5,10 +5,6 @@ All notable changes to this repository will be documented in this file.
 Format and process
 
 - The `Unreleased` section contains completed tasks that have been removed from `TODO.md` and are targeted for the next release.
-- integration.test.task.003 — undefined — Test cleanup
-- integration.test.task.002 — undefined — Integration test completion
-- integration.test.task.003 — undefined — Test cleanup
-- integration.test.task.002 — undefined — Integration test completion
 - Each entry should include: task id, summary, author, PR or commit link, and a one-line description of the change.
 - When creating a release, move the entries from `Unreleased` to a new versioned section (e.g. `## [1.0.0] - 2025-09-20`) and include release notes.
