+++ src/commands/shared/command.factory.ts
@@ -1,17 +1,16 @@
 import { Command } from 'commander';
 
+import { ILogger, IOutputWriter } from '../../types';
+import { RefAuditCommand } from '../audit/ref-audit.command';
+import { SupersedeCommand } from '../audit/supersede.command';
+import { NextCommand } from '../rendering/next.command';
+import { RenderCommand } from '../rendering/render.command';
 import { AddTaskCommand } from '../task-management/add-task.command';
 import { CompleteTaskCommand } from '../task-management/complete-task.command';
 import { ListTasksCommand } from '../task-management/list-tasks.command';
 import { ShowTaskCommand } from '../task-management/show-task.command';
 import { ValidateAndFixCommand } from '../validation/validate-and-fix.command';
 import { ValidateTasksCommand } from '../validation/validate-tasks.command';
-import { NextCommand } from '../rendering/next.command';
-import { RenderCommand } from '../rendering/render.command';
-import { RefAuditCommand } from '../audit/ref-audit.command';
-import { SupersedeCommand } from '../audit/supersede.command';
-import { ILogger } from '../../types/observability';
-import { IOutputWriter } from '../../types/rendering';
 
 /**
  * Factory for creating and configuring CLI commands following Command pattern.
