+++ src/commands/task-management/show-task.command.ts
@@ -1,25 +1,17 @@
 import { Command } from 'commander';
 
-import { ILogger } from '../../types/observability';
-import { TaskManager } from '../../core/storage/task.manager';
-import { EXIT_CODES } from '../../constants/exit-codes';
+import { ConsoleOutputWriter } from '../../core/rendering';
+import { TaskManager } from '../../core/storage';
+import {
+  CommandName,
+  ILogger,
+  IOutputWriter,
+  TodoShowCommandArgs,
+  EXIT_CODES,
+  TaskDetails,
+  OutputFormat,
+} from '../../types';
 import { BaseCommand } from '../shared/base.command';
-import { CommandName } from '../../types';
-import { IOutputWriter } from '../../types/rendering';
-import { ConsoleOutputWriter } from '../../core/rendering/console-output.writer';
-
-interface TaskDetails {
-  detailed_requirements?: unknown;
-  validations?: unknown;
-}
-
-/**
- * Arguments for the 'todo show' command
- */
-interface TodoShowCommandArgs {
-  /** Task ID to show */
-  id: string;
-}
 
 /**
  * Modern command for showing detailed information about a specific task.
@@ -61,7 +53,10 @@ export class ShowTaskCommand extends BaseCommand {
     this.outputWriter.newline();
 
     try {
-      this.outputWriter.writeFormatted((task as TaskDetails).detailed_requirements ?? {}, 'json');
+      this.outputWriter.writeFormatted(
+        (task as TaskDetails).detailed_requirements ?? {},
+        OutputFormat.JSON,
+      );
     } catch {
       this.outputWriter.warning('(invalid or missing detailed_requirements)');
       this.logger.warn('Invalid detailed_requirements in task', { id: args.id });
@@ -71,7 +66,7 @@ export class ShowTaskCommand extends BaseCommand {
     this.outputWriter.write('Validations:');
     this.outputWriter.newline();
     try {
-      this.outputWriter.writeFormatted((task as TaskDetails).validations ?? {}, 'json');
+      this.outputWriter.writeFormatted((task as TaskDetails).validations ?? {}, OutputFormat.JSON);
     } catch {
       this.outputWriter.warning('(invalid or missing validations)');
       this.logger.warn('Invalid validations in task', { id: args.id });
