+++ src/core/processing/hydrate.ts
@@ -7,8 +7,7 @@ import type {
   ITaskHydrationUseCase,
 } from '../../types/tasks';
 import { ILogger, TaskProviderType } from '../../types';
-import { UidStatusError } from '../../errors/uid-status.error';
-import { UidResolutionError } from '../../errors/uid-resolution.error';
+import { UidStatusError, UidResolutionError } from '../../types/core';
 import { Resolver } from '../helpers/uid-resolver';
 import { Renderer } from '../rendering/renderer';
 import { TaskProviderFactory } from '../storage/task-provider.factory';
