+++ src/types/commands/audit/index.ts
@@ -1,0 +1,8 @@
+/**
+ * Audit command types
+ */
+
+export interface ISupersedeOptions {
+  oldUid: string;
+  newUid: string;
+}
