+++ src/types/commands/index.ts
@@ -1,2 +1,7 @@
 export type { ICommand } from './ICommand';
 export { CommandName } from './CommandName';
+
+// Command-specific types
+export * from './task-management';
+export * from './rendering';
+export * from './audit';
