+++ src/types/commands/rendering/index.ts
@@ -1,0 +1,11 @@
+import type { IObservabilityLogger } from '../../observability';
+
+/**
+ * Rendering command types
+ */
+
+export interface OperationContext {
+  operationLogger: IObservabilityLogger;
+  startTime: Date;
+  stopTimer: () => void;
+}
