+++ src/types/commands/task-management/index.ts
@@ -1,0 +1,13 @@
+/**
+ * Task management command types
+ */
+
+export interface TaskDetails {
+  detailed_requirements?: unknown;
+  validations?: unknown;
+}
+
+export interface TodoShowCommandArgs {
+  /** Task ID to show */
+  id: string;
+}
