+++ src/types/core/index.ts
@@ -3,3 +3,7 @@ export * from './IFileManager';
 export * from './IServiceRegistry';
 export * from './IServiceResolver';
 export * from './SERVICE_KEYS';
+export * from './exit-codes';
+export * from './errors';
+export * from './registry';
+export * from './parsers';
