+++ src/types/core/parsers.ts
@@ -1,0 +1,15 @@
+import type { ILogger } from '../observability';
+
+import type { IFileManager } from './IFileManager';
+
+/**
+ * Parser-related types for file processing operations
+ */
+
+export interface UpdateYamlBlockOptions {
+  filePath: string;
+  id: string;
+  updatedData: Record<string, unknown>;
+  fileSystem: IFileManager;
+  logger?: ILogger;
+}
