+++ src/types/repository/index.ts
@@ -1,3 +1,4 @@
 export * from './ITaskRepository';
 export * from './IResolver';
 export * from './IExclusionFilter';
+export * from './github';
